05-high-low-gear: move domain tests to service tests.
This has the advantage to expose the feature test of the service layer
only to public parts of the domain. So if we want to change the domain,
we can wo changing a lot of tests. So diminishing the number of test on
the model allows for ore flexibility. But we still have good coverage of
the features on the service side.

Moreover, by moving those tests, we can decouple the service layer tests from
the model by interacting trough the repo.